gst/msdk: move AV1 encoder test classes to lib
